+++ src/main/java/error/InvalidSheep.java
@@ -1,0 +1,7 @@
+package error;
+
+public class InvalidSheep extends Exception {
+  public InvalidSheep (String message) {
+    super (message);
+  }
+}