+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 *Simple katas Changelog.*
 
+0.1.0
+- Added sheep counter kata.
+- Simple refactoring of error classes into a package.
+
 0.0.2
 - Added in some conditions for verifying input to the phone number creator
 (conditions not required by the codewars exercise, but should have been).